refactor: move codes in the scheme/alert package to scheme package
