--- internal/scheme/alert/doc.go
@@ -1,2 +1,0 @@
-// alert package implements how to kick the alert.
-package alert
