+++ internal/scheme/doc.go
@@ -1,2 +1,2 @@
-// probe package implements how to check the target status.
+// scheme package implements how to probe or send alert to the URL.
 package scheme
