+++ internal/scheme/testdata/ayd-bar-probe.bat
