+++ internal/scheme/testdata/ayd-broken-alert.bat
