+++ internal/scheme/testdata/ayd-foo-alert.bat
